requirements.txt
flask==3.1.1
pyyaml==6.0.2